Move CfdOrder inside projection actor
